Life status updates
Updated game to update eggCount for each snake to 0 when they die.

Updated game to send message to controller when snake is spawned to alive (1)

Updated game to send messages to controller when a snake dies in snakeFuneral. That snake's egg count is now 0 and the snakeLifeStatus is now 0. Updates the actual eggCount in the game to 0.